Separate duplex jobs + select inbox job
When duplex printing a full "vanilla" inbox, each inbox job is printed
as a separate job.

All inbox jobs are concatenated and handled as one virtual document that
is the object of PDF creation and proxy printing. However, when inbox is
"vanilla", user can select a single inbox job for PDF creation and proxy
printing.

Resolves: #541, #544